+++ python/ray/data/_internal/datasource/json_datasink.py
@@ -2,8 +2,8 @@ from typing import Any, Callable, Dict, Optional
 
 import pyarrow
 
+from ray.data._internal.util import _resolve_kwargs
 from ray.data.block import BlockAccessor
-from ray.data.datasource.file_based_datasource import _resolve_kwargs
 from ray.data.datasource.file_datasink import BlockBasedFileDatasink
