+++ python/ray/data/_internal/datasource/parquet_datasink.py
@@ -4,9 +4,8 @@ from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, List, Optional
 
 from ray.data._internal.arrow_ops.transform_pyarrow import concat
 from ray.data._internal.execution.interfaces import TaskContext
-from ray.data._internal.util import call_with_retry
+from ray.data._internal.util import _resolve_kwargs, call_with_retry
 from ray.data.block import Block, BlockAccessor
-from ray.data.datasource.file_based_datasource import _resolve_kwargs
 from ray.data.datasource.file_datasink import _FileDatasink
 from ray.data.datasource.filename_provider import FilenameProvider
